+++ threedigrid/orm/fields.py
@@ -15,13 +15,8 @@ from threedigrid.geo_utils import select_points_by_tile, \
     select_geoms_by_geometry, raise_import_exception
 from threedigrid.geo_utils import select_lines_by_tile
 from threedigrid.geo_utils import select_points_by_bbox
-from threedigrid.geo_utils import transform_xys
+from threedigrid.geo_utils import transform_xys, get_transformer
 from six.moves import map
-
-try:
-    from pyproj import Transformer
-except ImportError:
-    Transformer = None
 
 try:
     import shapely
@@ -212,28 +207,11 @@ class MultiLineArrayField(GeomArrayField):
         """
         reshaped_values = list(map(reshape_flat_array, values))
 
-        # Pyproj has (a lot) more overhead
-        # from 2.0.1 for reprojecting.
-        # so created the Transform once
-        # and reuse it.
-        if Transformer is not None:
-            # Only use the Transformer if it is present
-            # in Python 3 it works well, however in Python 2.7
-            # it seams to break
-            transformer = Transformer.from_proj(
-                int(source_epsg), int(target_epsg), always_xy=True
-            )
-
-            transform_values = [
-                np.array(transformer.transform(x[0], x[1])).flatten()
-                for x in reshaped_values
-            ]
-        else:
-            transform_values = [
-                transform_xys(x[0], x[1], source_epsg, target_epsg).flatten()
-                for x in reshaped_values
-            ]
-
+        transformer = get_transformer(source_epsg, target_epsg)
+        transform_values = [
+            np.array(transformer.transform(x[0], x[1])).flatten()
+            for x in reshaped_values
+        ]
         return np.array(transform_values, dtype=object)
 
     def _to_shapely_geom(self, values):